nivelmeter component created
